@@ -1,10 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
-using GuIdentitySample.Mvc.Models;
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
 using Msmaldi.AspNetCore.GuIdentity.EntityFrameworkCore;
+using GuIdentitySample.Mvc.Models;
 
 namespace GuIdentitySample.Mvc.Data
 {
