@@ -3,6 +3,7 @@ using Msmaldi.AspNetCore.GuIdentity;
 namespace GuIdentitySample.Mvc.Models
 {
     public class User : GuIdentityUser
-    {        
+    {
+        
     }
 }
