@@ -1,6 +1,6 @@
 ﻿@{
     ViewData["Title"] = "Privacy Policy";
 }
-<h2>@ViewData["Title"]</h2>
+<h1>@ViewData["Title"]</h1>
 
 <p>Use this page to detail your site's privacy policy.</p>
