@@ -1,6 +1,7 @@
 using Xunit;
 using System;
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
+using Microsoft.AspNetCore.Builder;
 using System.Reflection;
 using Microsoft.Extensions.DependencyInjection;
 using Msmaldi.AspNetCore.GuIdentity.EntityFrameworkCore;
@@ -14,15 +15,20 @@ namespace Msmaldi.AspNetCore.GuIdentity.EntityFrameworkCore.Test
 		public void EnsureGetManagers()
 		{
 			var services = new ServiceCollection();
+
+			services.AddLogging(config => { });
+
 			services.AddDbContext<InMemoryGuIdentityDbContext>();
+
 			services.AddIdentity<GuIdentityUser, GuIdentityRole>()
-					.AddEntityFrameworkGuidStores<InMemoryGuIdentityDbContext>()
-                	.AddDefaultTokenProviders();
+			 	.AddEntityFrameworkGuidStores<InMemoryGuIdentityDbContext>()
+        	   	.AddDefaultTokenProviders();
+
 			var userManager = services.BuildServiceProvider()
-							   		.GetRequiredService<UserManager<GuIdentityUser>>();
+				.GetRequiredService<UserManager<GuIdentityUser>>();
 
 			var signInManager = services.BuildServiceProvider()
-									.GetRequiredService<SignInManager<GuIdentityUser>>();
+			    .GetRequiredService<SignInManager<GuIdentityUser>>();
 			
 			Assert.NotNull(userManager);
 			Assert.NotNull(signInManager);			
